refactor rezolus sampler
Moves the `Statistic` enum out into its own module
